+++ projects/quang-core/src/lib/input-text/input-text.component.ts
@@ -9,156 +9,168 @@ import {
   Renderer2,
   Self,
   SimpleChanges,
-  ViewChild
-} from '@angular/core'
-import { ControlValueAccessor, NgControl } from '@angular/forms'
-import { delay, filter } from 'rxjs/operators'
+  ViewChild,
+} from "@angular/core";
+import { ControlValueAccessor, NgControl } from "@angular/forms";
+import { delay, filter } from "rxjs/operators";
 
 /**
  * input text component decorator
  */
 @Component({
-  selector: 'quang-input-text',
-  templateUrl: './input-text.component.html',
+  selector: "quang-input-text",
+  templateUrl: "./input-text.component.html",
   styles: [],
 })
 /**
  * input text component
  */
-export class InputTextComponent implements ControlValueAccessor, AfterViewInit, OnChanges, OnInit {
+export class InputTextComponent
+  implements ControlValueAccessor, AfterViewInit, OnChanges, OnInit
+{
   /**
    * Html id of input
    */
-  @Input() id: string = ''
+  @Input() id: string = "";
   /**
    * The label to display on the input field
    */
-  @Input() label: string = ''
+  @Input() label: string = "";
   /**
    * The placeholder of the input field
    */
-  @Input() placeholder: string = ''
+  @Input() placeholder: string = "";
   /**
    * The name of the form, this input is used to create keys for error, validation or help messages.
    * It will be the first key element generated
    */
-  @Input() formName: string = ''
+  @Input() formName: string = "";
   /**
    * Determine the arialabel tag for accessibility,
    * If not specified, it takes 'input' concatenated to the label by default
    */
-  @Input() ariaLabel: string = `Input ${this.label}`
+  @Input() ariaLabel: string = `Input ${this.label}`;
   /**
    * Array of additional classes to the input field
    */
-  @Input() customClass: string[] = []
+  @Input() customClass: string[] = [];
   /**
    * Defines if you want to display the help message for the user
    */
-  @Input() helpMessage: boolean = false
+  @Input() helpMessage: boolean = false;
 
   /**
    * Defines if you want to display the error message for the user
    */
-  @Input() errorMessage: boolean = false
+  @Input() errorMessage: boolean = false;
   /**
    * Defines if you want to display the success message for the user
    */
-  @Input() successMessage: boolean = false
+  @Input() successMessage: boolean = false;
   /**
    * Indicate the position in the page navigation flow with the tab key
    */
-  @Input() tabIndex: number = 0
+  @Input() tabIndex: number = 0;
   /**
    * Defines the autocomplete tag to indicate to the browser what type of field it is
    * and how to help the user fill it in
    */
-  @Input() autocomplete: string = 'off'
+  @Input() autocomplete: string = "off";
   /**
    * Defines the minimum length of the input field
    */
-  @Input() min: number = 0
+  @Input() min: number = 0;
   /**
    * Defines the maximum length of the input field
    */
-  @Input() max: number = 0
+  @Input() max: number = 0;
   /**
    * Indicates whether, when the page is opened,
    * this input field should be displayed in a focused state or not
    */
-  @Input() autofocus: boolean = false
+  @Input() autofocus: boolean = false;
   /**
    * Defines the validation pattern of the input field
    */
-  @Input() pattern: string = ''
+  @Input() pattern: string = "";
   /**
    * Defines whether the input field is in a read-only state
    */
-  @Input() readonly: boolean = false
+  @Input() readonly: boolean = false;
   /**
    * Adds bootstrap classes to the input that define the size of the field,
    * if not specified the field is displayed with standard size
    */
-  @Input() size: 'lg' | 'sm' | null = null
+  @Input() size: "lg" | "sm" | null = null;
+  /**
+   * set default start value in select option
+   */
+  @Input() defaultValue: string = "";
+
   /**
    * The value of the input
    */
-  _value: string = ''
+  _value: string = "";
   /**
    * the status of the success message
    */
-  _successMessage: string = ''
+  _successMessage: string = "";
   /**
    * the status of the error message
    */
-  _errorMessage: string = ''
+  _errorMessage: string = "";
   /**
    * the status of the help message
    */
-  _helpMessage: string = ''
+  _helpMessage: string = "";
   /**
    * Contains the value required by a validation when it fails
    */
-  _requiredValue: any = ''
+  _requiredValue: any = "";
   /**
    * Define disabled state
    */
-  _disabled: boolean = false
+  _disabled: boolean = false;
   /**
    * Standard definition to create a control value accessor
    */
-  onTouched(){}
+  onTouched() {}
   /**
    * Standard definition to create a control value accessor
    */
-  onChanged(value: string){}
+  onChanged(value: string) {}
 
   /**
    * The html input element
    */
-  @ViewChild('input', { static: true }) input: ElementRef<HTMLInputElement> | null = null
+  @ViewChild("input", { static: true })
+  input: ElementRef<HTMLInputElement> | null = null;
 
   /**
    * constructor
    * @param renderer html access
    * @param control cva access
    */
-  constructor (
+  constructor(
     private readonly renderer: Renderer2,
     @Self() @Optional() public control: NgControl
   ) {
-    this.control.valueAccessor = this
+    this.control.valueAccessor = this;
   }
 
   /**
    * Check if the help message is required and create the key
    */
-  ngOnInit (): void {
+  ngOnInit(): void {
     if (this.helpMessage) {
-      this._helpMessage = `${this.formName}.${this.control?.name}.help`
+      this._helpMessage = `${this.formName}.${this.control?.name}.help`;
     }
     if (this.successMessage) {
-      this._successMessage = `${this.formName}.${this.control?.name}.valid`
+      this._successMessage = `${this.formName}.${this.control?.name}.valid`;
+    }
+    if (this.defaultValue) {
+      this.writeValue(this.defaultValue);
+      this.onChanged(this._value);
     }
   }
 
@@ -166,37 +178,37 @@ export class InputTextComponent implements ControlValueAccessor, AfterViewInit, 
    * Checks if focus is required when displaying the input field.
    * Start the check on the validation of the field
    */
-  ngAfterViewInit (): void {
+  ngAfterViewInit(): void {
     setTimeout(() => {
       if (this.autofocus) {
-        this.input?.nativeElement.focus()
+        this.input?.nativeElement.focus();
       }
-    }, 0)
-    this.observeValidate()
+    }, 0);
+    this.observeValidate();
   }
 
   /**
    * Checks if focus is required when displaying the input field.
    * @param changes component changes
    */
-  ngOnChanges (changes: SimpleChanges): void {
+  ngOnChanges(changes: SimpleChanges): void {
     if (changes.autofocus?.currentValue && this.input) {
-      this.input.nativeElement.focus()
+      this.input.nativeElement.focus();
     }
   }
 
   /**
    * Standard definition to create a control value accessor
    */
-  registerOnTouched (fn: any): void {
-    this.onTouched = fn
+  registerOnTouched(fn: any): void {
+    this.onTouched = fn;
   }
 
   /**
    * Standard definition to create a control value accessor
    */
-  registerOnChange (fn: any): void {
-    this.onChanged = fn
+  registerOnChange(fn: any): void {
+    this.onChanged = fn;
   }
 
   /**
@@ -204,28 +216,32 @@ export class InputTextComponent implements ControlValueAccessor, AfterViewInit, 
    * its value is retrieved from the html element and the status change is signaled to the form
    * @param e
    */
-  onChangedHandler (e: Event): void {
-    this._value = (e.target as HTMLInputElement).value
-    this.onTouched()
-    this.onChanged(this._value)
+  onChangedHandler(e: Event): void {
+    this._value = (e.target as HTMLInputElement).value;
+    this.onTouched();
+    this.onChanged(this._value);
   }
 
   /**
    * Standard definition to create a control value accessor
    * When the value of the input field from the form is set, the value of the input html tag is changed
    */
-  writeValue (value: string): void {
-    this._value = value
-    this.renderer.setProperty(this.input?.nativeElement, 'value', value)
+  writeValue(value: string): void {
+    this._value = value;
+    this.renderer.setProperty(this.input?.nativeElement, "value", value);
   }
 
   /**
    * Standard definition to create a control value accessor
    * When the input field from the form is disabled, the html input tag is defined as disabled
    */
-  setDisabledState (isDisabled: boolean): void {
-    this.renderer.setProperty(this.input?.nativeElement, 'disabled', isDisabled)
-    this._disabled = isDisabled
+  setDisabledState(isDisabled: boolean): void {
+    this.renderer.setProperty(
+      this.input?.nativeElement,
+      "disabled",
+      isDisabled
+    );
+    this._disabled = isDisabled;
   }
 
   /**
@@ -234,23 +250,25 @@ export class InputTextComponent implements ControlValueAccessor, AfterViewInit, 
    * If there is an error with a specific required value it is passed to the translation pipe
    * to allow for the creation of custom messages
    */
-  observeValidate (): void {
-    this.control?.statusChanges?.pipe(
-      delay(0),
-      filter(() => !!this.control.dirty)
-    ).subscribe(() => {
-      if (this.control.invalid && this.errorMessage) {
-        if (this.control.errors) {
-          for (const error in this.control.errors) {
-            if (error === 'minlength' || error === 'maxlength') {
-              this._requiredValue = this.control.errors[error].requiredLength
-            } else {
-              this._requiredValue = this.control.errors[error].requiredValue
+  observeValidate(): void {
+    this.control?.statusChanges
+      ?.pipe(
+        delay(0),
+        filter(() => !!this.control.dirty)
+      )
+      .subscribe(() => {
+        if (this.control.invalid && this.errorMessage) {
+          if (this.control.errors) {
+            for (const error in this.control.errors) {
+              if (error === "minlength" || error === "maxlength") {
+                this._requiredValue = this.control.errors[error].requiredLength;
+              } else {
+                this._requiredValue = this.control.errors[error].requiredValue;
+              }
+              this._errorMessage = `${this.formName}.${this.control?.name}.${error}`;
             }
-            this._errorMessage = `${this.formName}.${this.control?.name}.${error}`
           }
         }
-      }
-    })
+      });
   }
 }